% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_net_model.R
\name{assemble_net_model}
\alias{assemble_net_model}
\title{Assemble bathtub inundation model}
\usage{
assemble_net_model(
  pipes,
  structures,
  type = "elevation",
  use_raster_elevation = T,
  elev,
  elev_units = "m",
  min_elev_cutoff = -5,
  buffer = 0.5,
  interp_rnds = 5,
  interp_adjust = 0.1,
  guess_connectivity = T,
  up_condition = NULL,
  dn_condition = NULL,
  structure_condition = NULL,
  obstruction_keywords = NULL,
  obstruction_percent = F,
  overwrite = T,
  site_name,
  workspace
)
}
\arguments{
\item{pipes}{A \code{sf} object denoting stormwater pipes. Result of \code{setup_pipes}}

\item{structures}{A \code{sf} object denoting stormwater structures. Result of \code{setup_structures}}

\item{type}{Type of invert survey data. 'depth', 'elevation', or 'none'}

\item{use_raster_elevation}{Extract elevation from DEM and use as surface elevations?}

\item{elev}{\code{RasterLayer} DEM object for inundation modeling. Result of \code{DEM_setup}}

\item{elev_units}{Units of \code{elev} values}

\item{min_elev_cutoff}{Minimum cutoff for elevation values}

\item{buffer}{Size of buffer for connecting structures to pipes}

\item{interp_rnds}{Rounds of interpolation for missing invert elevations}

\item{interp_adjust}{Elevation value to add to nearby known elevation to estimate unknowns. Default is 0.1}

\item{guess_connectivity}{Assume connectivity of pipes that overlap without nodes?}

\item{up_condition}{Column name of upstream pipe condition. Either text of condition to be
searched with \code{obstruction_keywords} or percent obstructions (\code{obstruction_percent})}

\item{dn_condition}{Column name of downstream pipe condition. Either text of condition to be
searched with \code{obstruction_keywords} or percent obstructions (\code{obstruction_percent})}

\item{structure_condition}{Column name of structure condition. Either text of condition to be
searched with \code{obstruction_keywords} or percent obstructions (\code{obstruction_percent})}

\item{obstruction_keywords}{Vector of keywords (in quotes) denoting obstruction to evaluate in
\code{up_condition}, \code{dn_condition}, \code{structure_condition}}

\item{obstruction_percent}{Should \code{up_condition}, \code{dn_condition}, and \code{structure_condition},
be evaluated as percent obstructed?}

\item{overwrite}{Overwrite existing model files in the workspace? Default is TRUE. If set to FALSE,
model will not be saved in output folder and will exist in memory (see load_model to load existing model).}

\item{workspace}{bathtub workspace path}
}
\value{
An list of three \code{sf} objects denoting Pipes, Nodes, and Structures. Files are saved in '/model/'
folder in workspace and by default overwrite existing model files
}
\description{
Assemble bathtub inundation model
}
\examples{
# bft_model <- assemble_net_model(
# pipes = pipes_n,
# structures = structures_n,
# type = "none",
# elev = bft_elev,
# elev_units = "m",
# use_raster_elevation = F,
# buffer = 1,
# guess_connectivity = T
# )
}
