% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_DEM.R
\name{DEM_setup}
\alias{DEM_setup}
\title{Setup DEM for inundation modeling}
\usage{
DEM_setup(
  pipes,
  other_network_layers = NULL,
  large_DEM,
  elev_units = "m",
  conversion_raster,
  res,
  trace_upstream = F,
  trace_buffer = 1.5,
  extent_buffer = 1000,
  minimum_area = 2,
  workspace,
  overwrite = F
)
}
\arguments{
\item{pipes}{A \code{sf} object denoting stormwater pipes.}

\item{large_DEM}{\code{RasterLayer} DEM to be processed}

\item{conversion_raster}{\code{RasterLayer} Conversion from NAVD88 to MHHW}

\item{res}{Desired spatial resolution of output DEM (meters)}

\item{trace_upstream}{Logical. Trace NAVD88/MHHW conversion factor upstream?}

\item{trace_buffer}{Buffer distance (map units of pipes) around pipes
to search for receiving waters}

\item{minimum_area}{km^2. Minimum area of flooding to keep as "downstream flooding"}

\item{workspace}{Path to bathtub folder}

\item{overwrite}{Logical. Overwrite existing data?}

\item{other_network_layer}{Vector of \code{sf} objects to be included in network}
}
\value{
DEM clipped to stormwater network, reprojected to projection
of \code{pipes}, and values converted from NAVD88 to MHHW
}
\description{
Setup DEM for inundation modeling
}
\examples{
# site_DEM <- bathtub::DEM_setup(
# pipes = site_pipes,
# large_DEM = NC_DEM,
# conversion_raster = NAVD88_MHHW_conv,
# res = 5,
# workspace = site_workspace
# )
}
