% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_folder.R
\name{folder_setup}
\alias{folder_setup}
\title{Create folders for bathtub model and results}
\usage{
folder_setup(path)
}
\arguments{
\item{path}{Full path for location of bathtub output folder}
}
\value{
Workspace path for bathtub models
}
\description{
Create folders for bathtub model and results
}
\examples{
# workspace <- folder_setup("your_path_here")
}
