% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_3DEP_elevation.R
\name{get_3DEP_elevation}
\alias{get_3DEP_elevation}
\title{Download USGS 3DEP DEM (1/9 arc second ~ 10 m)}
\usage{
get_3DEP_elevation(x, topobathy = F, view = F)
}
\arguments{
\item{x}{A \code{sf} object denoting stormwater pipes used to calculate extent.}

\item{topobathy}{Select only topographic+bathymetric data?}

\item{view}{View available data tile extent to assist in selecting data?}
}
\value{
USGS 3DEP 1/9 arc-second DEM clipped to stormwater network extent, reprojected to projection
of \code{x}.
}
\description{
Download USGS 3DEP DEM (1/9 arc second ~ 10 m)
}
\examples{
usgs_elev <- get_3DEP_elevation(x = pipes, topobathy = T, view = T)
}
