% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NOAA_elevation.R
\name{get_NOAA_elevation}
\alias{get_NOAA_elevation}
\title{Download NOAA SLR DEM}
\usage{
get_NOAA_elevation(
  x,
  x_EPSG = 102719,
  DEM_EPSG = x_EPSG,
  res = 9.81,
  min_elev_cutoff = -5,
  workspace
)
}
\arguments{
\item{x}{A \code{sf} object denoting stormwater pipes used to calculate extent}

\item{x_EPSG}{EPSG code of \code{x}}

\item{DEM_EPSG}{EPSG code of returned DEM. Default is same as \code{x_EPSG}}

\item{res}{Desired resolution of DEM in map units of \code{x}}

\item{min_elev_cutoff}{Minimum elevation to keep (meters). NOAA SLR DEM uses -99 to denote
permanent water surfaces}

\item{workspace}{Path to floodr folder}
}
\value{
NOAA SLR DEM clipped to stormwater network extent, reprojected to projection
of \code{x}
}
\description{
Download NOAA SLR DEM
}
\examples{
noaa_elev <- get_NOAA_elevation(x = pipes, workspace = workspace)
}
