% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{Load existing bathtub model from file}
\usage{
load_model(workspace)
}
\arguments{
\item{workspace}{Full path for location of bathtub output folder}
}
\value{
bathtub model
}
\description{
Load existing bathtub model from file
}
\examples{
# bft_model <- load_model(workspace)
}
