% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_results.R
\name{load_results}
\alias{load_results}
\title{Load existing bathtub model results from file}
\usage{
load_results(workspace)
}
\arguments{
\item{workspace}{Full path for location of bathtub output folder}
}
\value{
bathtub model results
}
\description{
Load existing bathtub model results from file
}
\examples{
# bft_model_results <- load_results(workspace)
}
