% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_w_units.R
\name{load_w_units}
\alias{load_w_units}
\title{Load sf files with units}
\usage{
load_w_units(x)
}
\arguments{
\item{x}{Full path for location of \code{sf} file saved with \code{save_w_units}}
}
\value{
A \code{sf} object with units from associated file. See \code{save_w_units} for more info
}
\description{
Load sf files with units
}
\examples{
# x_units <- load_w_units("your_path_here")
}
