% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAVD88_MHHW_NC.R
\docType{data}
\name{NAVD88_MHHW_NC}
\alias{NAVD88_MHHW_NC}
\title{Raster with conversion factor from NAVD88 to MHHW for North Carolina, USA.}
\format{
RasterLayer:
\describe{
  \item{dimensions}{6197 X 6775 pixels}
  \item{resolution}{50 meter}
  \item{crs/projection}{UTM zone 18}
  \item{value range}{-0.99, 0.7606}
  ...
}
}
\source{
\url{https://www.fisheries.noaa.gov/inport/item/48104}
}
\usage{
NAVD88_MHHW_NC
}
\description{
Values are in meters. To convert from NAVD88 to MHHW tidal datum,
subtract the conversion raster from a DEM with NAVD88 elevation values.
}
\keyword{datasets}
