% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_w_units.R
\name{save_w_units}
\alias{save_w_units}
\title{Write sf files with units}
\usage{
save_w_units(x, full_path, overwrite = T)
}
\arguments{
\item{x}{\code{sf} object with units.}

\item{full_path}{Full path (including filename and extension) to write \code{sf} object.
CSV file with column names and units will be saved with the same path but with
"_units.csv" at the end.}

\item{overwrite}{Overwrite existing file? Default is T}
}
\value{
N/A
}
\description{
Write sf files with units
}
\examples{
# save_w_units(
# x = pipes,
# full_path = "your_path_here/pipes.gpkg",
# overwrite = T)
}
