% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_pipes.R
\name{setup_pipes}
\alias{setup_pipes}
\title{Setup pipes for floodr inundation modeling}
\usage{
setup_pipes(
  pipes,
  type = "depth",
  up_invert,
  dn_invert,
  null_value = 99,
  invert_units = "ft",
  up_elev = NA,
  dn_elev = NA,
  elev_units = "ft",
  vdatum = "NAVD88",
  diam,
  diam_units,
  other_cols = NULL,
  workspace
)
}
\arguments{
\item{pipes}{A \code{sf} object denoting stormwater pipes}

\item{type}{Type of invert survey data. 'depth', 'elevation', or 'none'}

\item{up_invert}{Name of column denoting upstream invert survey data}

\item{dn_invert}{Name of column denoting downstream invert survey data}

\item{null_value}{Invert survey data value indicating missing data}

\item{up_elev}{Name of column denoting surface elevation of upstream pipe ends}

\item{dn_elev}{Name of column denoting surface elevation of downstream pipe ends}

\item{elev_units}{Units of \code{up/dn_invert} and \code{up/dn_elev}}

\item{vdatum}{Vertical datum of elevation data (only NAVD88 supported)}

\item{diam}{Name of column denoting pipe diameter}

\item{diam_units}{Units of \code{diam}}

\item{other_cols}{Other columns to preserve}

\item{workspace}{Path to floodr folder}
}
\value{
An \code{sf} object denoting pipes formatted with units
}
\description{
Setup pipes for floodr inundation modeling
}
\examples{
pipes_n <-
setup_pipes(pipes,
           type = "none",
           diam = "DIAMETER",
           diam_units = "in")
}
