% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_structures.R
\name{setup_structures}
\alias{setup_structures}
\title{Setup structures for bathtub inundation modeling}
\usage{
setup_structures(
  structures,
  type = "depth",
  invert,
  invert_units = "ft",
  elev = NA,
  elev_units = "m",
  vdatum = "NAVD88",
  null_value = 99,
  other_cols = NULL,
  workspace
)
}
\arguments{
\item{structures}{A \code{sf} object denoting stormwater structures}

\item{type}{Type of invert survey data. 'depth', 'elevation', or 'none'}

\item{invert}{Name of column denoting invert survey data}

\item{invert_units}{Units of \code{invert} elevation data}

\item{elev}{Name of column denoting surface elevation of stormwater structure,
or \code{RasterLayer} NAVD88 DEM to extract elevation from}

\item{elev_units}{Units of \code{elev} values}

\item{vdatum}{Vertical datum of elevation data (only NAVD88 supported)}

\item{null_value}{Invert survey data value indicating missing data}

\item{other_cols}{Other columns to preserve}

\item{workspace}{Path to bathtub folder}
}
\value{
An \code{sf} object denoting structures formatted with units
}
\description{
Setup structures for bathtub inundation modeling
}
\examples{
# structures_n <-
# setup_structures(
#   structures = structures,
#   type = "elevation",
#   invert = "INVERTELEV",
#   elev = "Elevation",
#   elev_units = "ft",
#   null_value = 99,
#   other_cols = c("Code","Prcnt_Obst","Type_Obst"),
#   workspace = workspace
# )
}
