% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_structures.R
\name{viz_structures}
\alias{viz_structures}
\title{Visualize impacts on structures from bathtub inundation modeling}
\usage{
viz_structures(
  model_output,
  model,
  elev,
  type_column = NULL,
  filter_value = NULL,
  type = "plot",
  hide_labels = F,
  simplify_labels = T,
  label_size = 2,
  xmin = -3,
  xmax = 4,
  filename = NULL,
  workspace,
  ...
)
}
\arguments{
\item{model_output}{A 'bathtub' model output item}

\item{model}{A 'bathtub' model item}

\item{elev}{The DEM used in modeling inundation created by 'setup_DEM'}

\item{type_column}{Column in model item that indicates the type of structures. Default is NULL}

\item{filter_value}{Type of structure (as string) to select from 'type_column'. Multiple values should be put in vector (i.e., using c() ). Default is NULL}

\item{type}{Static "plot" or "interactive_map"}

\item{hide_labels}{Hide labels of column values? T or F, default is F}

\item{simplify_labels}{Simplify labels of column values? T or F, default is T}

\item{label_size}{Size of labels, default is 2}

\item{xmin}{Set the minimum value of the x axis}

\item{xmax}{Set the maximum value of x axis}

\item{filename}{Name of the plot. Will be stored in the 'bathtub_output/figures/}

\item{workspace}{Path to bathtub output folder}
}
\value{
A \code{ggplot} plot (type = "plot") or \code{mapview} map (type = "interactive_map") of structures impacted during inundation modeling
}
\description{
Visualize impacts on structures from bathtub inundation modeling
}
\examples{
bft_plot <- viz_structures(model_output = bft_model_output,
model = bft_model_1D,
elev = raster::raster(bft_elev),
type_column = "Code",
filter_value = c("D_I", "C_B"),
type = "plot",
hide_labels = T,
simplify_labels = T,
label_size = 2,
filename = "bft_structures_impact_plot.png",
workspace = workspace,
panel.grid = element_blank())
}
